Rename next_level_trackqc.py
